requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
coverage==7.11.0
crispy-tailwind==1.0.3
Django==5.2.5
django-crispy-forms==2.5
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
flake8==7.3.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
parameterized==0.9.0
pathspec==0.12.1
pillow==11.3.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
validate_docbr==1.11.1
